requirements.txt
# Core ML and Data Science
torch>=1.9.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
seaborn>=0.11.0

# Explainability Libraries
shap>=0.41.0
dice-ml>=0.9.0

# Streamlit and UI
streamlit>=1.20.0

# LLM Integration
transformers>=4.20.0
huggingface_hub>=0.15.0
requests>=2.28.0

# Environment Management
python-dotenv>=0.19.0